Improved import_in_calcurse script - mutt

+++ configs/mutt/scripts/import_in_calcurse.sh
@@ -1,7 +1,8 @@
 #!/bin/dash
 aptsfile=/home/ashish/.local/share/calcurse/apts
 
-calcurse -i "$1"
-uniq "$aptsfile" /tmp/apts.temp
+read -r -p "Import event(s) in calcurse? [Y/n]: " confirm
+[ "$confirm" = n ] || [ "$confirm" = N ] && exit
+calcurse -i "$1" >/dev/null 2>&1
+awk '!x[$0] {x[$0]=1; print}' "$aptsfile" >/tmp/apts.temp
 mv /tmp/apts.temp "$aptsfile"
-notify-send -t 1000 Calcurse "Import successful"
